requirements.txt
fastapi==0.132.0
uvicorn[standard]==0.41.0
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
python-jose[cryptography]==3.5.0
python-multipart==0.0.22
pydantic[email]==2.12.5
python-dotenv==1.2.1
bcrypt==4.2.1
pymupdf==1.24.10
Pillow==10.4.0
pandas>=2.0.0
docling>=2.70.0
anthropic>=0.50.0
google-generativeai==0.8.5